requirements.txt
Flask
flask-mysqldb
python-dotenv
pyjwt